1.5.5 - Added AttributePanel for setting keyframes via different keyframe-specific cards - Added basic keyframe functionality - Updated KeyFrame classes to work with updated TimeSheet - Added timeSpinner to TimeSheetPanel - TimeSheet: added preview time indicator, added drag functionality, added keyframe drawing, moved mouseListeners

+++ src/main/java/com/creatorskit/swing/timesheet/keyframe/KeyFrame.java
@@ -10,5 +10,5 @@ import lombok.Setter;
 public class KeyFrame
 {
     private KeyFrameType keyFrameType;
-    private int tick;
+    private double tick;
 }
